This is easily the most awful thing I’ve ever done
