+++ src/main/java/org/tomitribe/util/hash/JvmUtils.java
@@ -13,58 +13,26 @@
  */
 package org.tomitribe.util.hash;
 
-import sun.misc.Unsafe;
+import static java.nio.NioPackageUtils.newDirectByteBuffer;
+import static sun.misc.Unsafe.getUnsafe;
 
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.MethodType;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
 import java.nio.ByteBuffer;
 
-import static org.tomitribe.util.hash.UnsafeConstants.*;
+import sun.misc.Unsafe;
 
 @SuppressWarnings("restriction")
 final class JvmUtils {
-    static final Unsafe unsafe;
-    static final MethodHandle newByteBuffer;
-
-    static {
-        try {
-            // fetch theUnsafe object
-            Field field = Unsafe.class.getDeclaredField("theUnsafe");
-            field.setAccessible(true);
-            unsafe = (Unsafe) field.get(null);
-            if (unsafe == null) {
-                throw new RuntimeException("Unsafe access not available");
-            }
-
-            // verify the stride of arrays matches the width of primitives
-            assertArrayIndexScale("Boolean", ARRAY_BOOLEAN_INDEX_SCALE, 1);
-            assertArrayIndexScale("Byte", ARRAY_BYTE_INDEX_SCALE, 1);
-            assertArrayIndexScale("Short", ARRAY_SHORT_INDEX_SCALE, 2);
-            assertArrayIndexScale("Int", ARRAY_INT_INDEX_SCALE, 4);
-            assertArrayIndexScale("Long", ARRAY_LONG_INDEX_SCALE, 8);
-            assertArrayIndexScale("Float", ARRAY_FLOAT_INDEX_SCALE, 4);
-            assertArrayIndexScale("Double", ARRAY_DOUBLE_INDEX_SCALE, 8);
-
-            // fetch a method handle for the hidden constructor for DirectByteBuffer
-            Class<?> directByteBufferClass = ClassLoader.getSystemClassLoader().loadClass("java.nio.DirectByteBuffer");
-            Constructor<?> constructor = directByteBufferClass.getDeclaredConstructor(long.class, int.class, Object.class);
-            constructor.setAccessible(true);
-            newByteBuffer = MethodHandles.lookup().unreflectConstructor(constructor)
-                    .asType(MethodType.methodType(ByteBuffer.class, long.class, int.class, Object.class));
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private static void assertArrayIndexScale(final String name, int actualIndexScale, int expectedIndexScale) {
-        if (actualIndexScale != expectedIndexScale) {
-            throw new IllegalStateException(name + " array index scale must be " + expectedIndexScale + ", but is " + actualIndexScale);
-        }
-    }
+    static final Unsafe unsafe = getUnsafe();
+    static final BackwardsCompatibleInnerClass newByteBuffer = new BackwardsCompatibleInnerClass();
 
     private JvmUtils() {
+    }
+
+    public static class BackwardsCompatibleInnerClass {
+
+        public ByteBuffer invokeExact(Object[] args) {
+            return newDirectByteBuffer(args);
+        }
+
     }
 }
